api/requirements.txt
Django==3.1.7
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0
celery==5.0.5
psycopg2-binary==2.8.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework-simplejwt==4.6.0
